# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Deltaswap Phylax API\nThis is the API for the Deltaswap Phylax and Explorer.\nThe API has two namespaces: deltaswapscan and phylax.\ndeltaswapscan is the namespace for the explorer and the new endpoints. The prefix is /api/v1.\nphylax is the legacy namespace backguard compatible with phylax node API. The prefix is /v1.\nThis API is public and does not require authentication although some endpoints are rate limited.\nCheck each endpoint documentation for more information.",
        "title": "Deltaswapscan API",
        "termsOfService": "https://wormhole.com/",
        "contact": {
            "name": "API Support",
            "url": "https://discord.com/invite/wormholecrypto",
            "email": "info@wormhole.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/signed_batch_vaa/{chain_id}/{emitter}/sequence/{seq}": {
            "get": {
                "description": "get a batch of VAA []byte from a chainID, emitter address and sequence.",
                "tags": [
                    "Phylax"
                ],
                "operationId": "phylaxs-find-signed-batch-vaa",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the blockchain",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "address of the emitter",
                        "name": "emitter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "sequence of the VAA",
                        "name": "seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "vaaBytes": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/signed_vaa/{chain_id}/{emitter}/{seq}": {
            "get": {
                "description": "get a VAA []byte from a chainID, emitter address and sequence.",
                "tags": [
                    "Phylax"
                ],
                "operationId": "phylaxs-find-signed-vaa",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the blockchain",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "address of the emitter",
                        "name": "emitter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "sequence of the VAA",
                        "name": "seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "vaaBytes": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
...
}